auth_project/requirements.txt
amqp==5.3.1
asgiref==3.11.1
async-timeout==5.0.1
billiard==4.2.4
celery==5.6.2
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==5.2.12
django-cors-headers==4.9.0
django-ratelimit==4.1.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
exceptiongroup==1.3.1
kombu==5.6.2
msgpack==1.1.2
packaging==26.0
prompt_toolkit==3.0.52
PyJWT==2.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
redis==7.2.1
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
vine==5.1.0
wcwidth==0.6.0